backend/pyproject.toml
[project]
name = "dailyai-backend"
version = "0.1.0"
description = "AI Chat Assistant for Calendar & Email"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "deepagents>=0.1",
    "langchain>=0.3",
    "langgraph>=0.4",
    "langchain-google-genai>=2.1",
    "google-api-python-client>=2.160",
    "google-auth-oauthlib>=1.2",
    "python-dotenv>=1.0",
]
